LPS-32433 Test Formatting
