+++ portal-web/test/functional/com/liferay/portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPortletCN3Test.html
