+++ portal-web/test/functional/com/liferay/portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPortletCNTests.html
@@ -10,6 +10,7 @@
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="AddPageCNTest.html">AddPageCNTest</a></td></tr>
 <tr><td><a href="AddPortletCNTest.html">AddPortletCNTest</a></td></tr>
+<tr><td><a href="ViewPortletCNTest.html">ViewPortletCNTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
