+++ portal-web/test/functional/com/liferay/portalweb/portlet/categoriesnavigation/portlet/addportletcn/ViewPortletCNTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.categoriesnavigation.portlet.addportletscn;
+package com.liferay.portalweb.portlet.categoriesnavigation.portlet.addportletcn;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,22 +20,23 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class ViewPortletsCNTest extends BaseTestCase {
-	public void testViewPortletsCN() throws Exception {
+public class ViewPortletCNTest extends BaseTestCase {
+	public void testViewPortletCN() throws Exception {
 		selenium.selectWindow("null");
 		selenium.selectFrame("relative=top");
 		selenium.open("/web/guest/home/");
 		selenium.clickAt("link=Categories Navigation Test Page",
 			RuntimeVariables.replace("Categories Navigation Test Page"));
 		selenium.waitForPageToLoad("30000");
-		assertTrue(selenium.isVisible("//div[1]/section"));
+		assertTrue(selenium.isVisible("//section"));
+		assertEquals(RuntimeVariables.replace("Categories Navigation"),
+			selenium.getText("//h1[@class='portlet-title']/span[2]"));
+		assertEquals(RuntimeVariables.replace("Options"),
+			selenium.getText("//span[@title='Options']/ul/li/strong/a"));
+		assertTrue(selenium.isVisible("//img[@title='Minimize']"));
+		assertTrue(selenium.isVisible("//img[@title='Minimize']"));
+		assertTrue(selenium.isVisible("//img[@title='Remove']"));
 		assertEquals(RuntimeVariables.replace("There are no categories."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[1]"));
-		assertTrue(selenium.isVisible("//div[2]/section"));
-		assertEquals(RuntimeVariables.replace("There are no categories."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[2]"));
-		assertTrue(selenium.isVisible("//div[3]/section"));
-		assertEquals(RuntimeVariables.replace("There are no categories."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[3]"));
+			selenium.getText("//div[@class='portlet-msg-info']"));
 	}
 }
