+++ portal-web/test/functional/com/liferay/portalweb/portlet/categoriesnavigation/portlet/addportletscn/AddPortletsCNTests.html
@@ -9,9 +9,9 @@
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPageCNTest.html">AddPageCNTest</a></td></tr>
-<tr><td><a href="AddPortletCN1Test.html">AddPortletCN1Test</a></td></tr>
-<tr><td><a href="AddPortletCN2Test.html">AddPortletCN2Test</a></td></tr>
-<tr><td><a href="AddPortletCN3Test.html">AddPortletCN3Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPortletCN1Test.html">AddPortletCN1Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPortletCN2Test.html">AddPortletCN2Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/categoriesnavigation/portlet/addportletcn/AddPortletCN3Test.html">AddPortletCN3Test</a></td></tr>
 <tr><td><a href="ViewPortletsCNTest.html">ViewPortletsCNTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
